data successfully fetched moving on to react app
