+++ firebaseConfig.js
@@ -1,0 +1,14 @@
+import { initializeApp } from "firebase/app";
+import { getFirestore, collection, getDocs } from "firebase/firestore"
+
+const firebaseConfig = {
+  apiKey: "AIzaSyDbDOEjx--5Wdsrlp7HnaSTV1rYo5TeAlo",
+  authDomain: "loafer-portfolio.firebaseapp.com",
+  projectId: "loafer-portfolio",
+  storageBucket: "loafer-portfolio.appspot.com",
+  messagingSenderId: "89801243994",
+  appId: "1:89801243994:web:a440cd1130f213e5ed8763"
+};
+
+export const app = initializeApp(firebaseConfig)
+export const database = getFirestore(app)
